requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pandas==2.2.3
numpy==2.2.2
requests==2.32.3
scikit-learn==1.6.1
python-dotenv==1.0.1
joblib==1.4.2
